tests/wpt/tests/css/css-color/parsing/color-computed-contrast-color-function.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Test CSS Color Module 5 contrast-color() computed value</title>
<link rel="author" title="Sam Weinig" href="mailto:sam@webkit.org">
<link rel="help" href="https://drafts.csswg.org/css-color-5/#contrast-color">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
<div id="target"></div>
<script>

// NOTE: As the algorithm for determining constrast is not defined, all computed values can validly be either `rgb(0, 0, 0)` or `rgb(255, 255, 255)`.

// Test basic legacy colors
test_computed_value("background-color", "contrast-color(white)", ["rgb(0, 0, 0)", "rgb(255, 255, 255)"]);
test_computed_value("background-color", "contrast-color(black)", ["rgb(0, 0, 0)", "rgb(255, 255, 255)"]);
test_computed_value("background-color", "contrast-color(pink)", ["rgb(0, 0, 0)", "rgb(255, 255, 255)"]);

// Test non-legacy colors
test_computed_value("background-color", "contrast-color(color(srgb 1 0 1 / 0.5))", ["rgb(0, 0, 0)", "rgb(255, 255, 255)"]);
test_computed_value("background-color", "contrast-color(lab(0.2 0.5 0.2))", ["rgb(0, 0, 0)", "rgb(255, 255, 255)"]);

// Test out-of-gamut colors
test_computed_value("background-color", "contrast-color(color(srgb 10 10 10))", ["rgb(0, 0, 0)", "rgb(255, 255, 255)"]);
test_computed_value("background-color", "contrast-color(color(srgb -10 -10 -10))", ["rgb(0, 0, 0)", "rgb(255, 255, 255)"]);

// Test nested contrast-color
test_computed_value("background-color", "contrast-color(contrast-color(pink))", ["rgb(0, 0, 0)", "rgb(255, 255, 255)"]);

// Test currentColor
test_computed_value("background-color", "contrast-color(currentcolor)", ["rgb(0, 0, 0)", "rgb(255, 255, 255)"]);

// Test color using calc().
test_computed_value("background-color", "contrast-color(color(srgb calc(1 + 1 / 1) calc(1 + 1 / 1) calc(1 + 1 / 1)))", ["rgb(0, 0, 0)", "rgb(255, 255, 255)"]);

// Test contrast-color in other color functions.
test_computed_value("background-color", "color-mix(in srgb, contrast-color(blue) 100%, purple)", ["color(srgb 0 0 0)", "color(srgb 1 1 1)"]);
test_computed_value("background-color", "rgb(from contrast-color(blue) r g b)", ["color(srgb 0 0 0)", "color(srgb 1 1 1)"]);

// Test color functions in contrast-color.
test_computed_value("background-color", "contrast-color(color-mix(blue, green))", ["rgb(0, 0, 0)", "rgb(255, 255, 255)"]);
test_computed_value("background-color", "contrast-color(rgb(from orange r g b))", ["rgb(0, 0, 0)", "rgb(255, 255, 255)"]);

// Test color using system color.
test_computed_value("background-color", "contrast-color(buttonface)", ["rgb(0, 0, 0)", "rgb(255, 255, 255)"]);

// Test color using alpha channel.
test_computed_value("background-color", "contrast-color(transparent)", ["rgb(0, 0, 0)", "rgb(255, 255, 255)"]);
test_computed_value("background-color", "contrast-color(rgba(255, 255, 255, 0.2))", ["rgb(0, 0, 0)", "rgb(255, 255, 255)"]);
</script>
